Per gevent/gevent#546, rebase the changes on top of gevent 1.0.2. Fixes gevent/gevent#477.
