+++ gevent/hub.py
@@ -26,17 +26,24 @@ __all__ = ['getcurrent',
            'Hub',
            'Waiter']
 
+# Sniff Python > 2.7.9 for new SSL interfaces
+# If True, Python is greater than or equal to 2.7.9 (but not Python 3).
+PYGTE279 = (
+    sys.version_info[0] == 2
+    and sys.version_info[1] >= 7
+    and sys.version_info[2] >= 9
+)
 
 PY3 = sys.version_info[0] >= 3
-
+PYPY = hasattr(sys, 'pypy_version_info')
 
 if PY3:
     string_types = str,
     integer_types = int,
+
 else:
     string_types = basestring,
     integer_types = (int, long)
-
 
 if sys.version_info[0] <= 2:
     import thread
@@ -279,7 +286,7 @@ class Hub(greenlet):
         else:
             try:
                 info = self.loop._format()
-            except Exception, ex:
+            except Exception as ex:
                 info = str(ex) or repr(ex) or 'error'
         result = '<%s at 0x%x %s' % (self.__class__.__name__, id(self), info)
         if self._resolver is not None:
