+++ greentest/test__all__.py
@@ -31,7 +31,7 @@ NOT_IMPLEMENTED = {
 COULD_BE_MISSING = {
     'socket': ['create_connection', 'RAND_add', 'RAND_egd', 'RAND_status']}
 
-NO_ALL = ['gevent.threading', 'gevent._util']
+NO_ALL = ['gevent.threading', 'gevent._util', 'gevent._socketcommon']
 
 
 class Test(unittest.TestCase):
@@ -138,6 +138,10 @@ are missing from %r:
             raise AssertionError(msg)
 
     def _test(self, modname):
+        if modname.endswith('2') or modname.endswith("279"):
+            return
+        if modname.endswith('3'):
+            return
         self.modname = modname
         exec "import %s" % modname in {}
         self.module = sys.modules[modname]
