+++ greentest/test__execmodules.py
@@ -1,3 +1,5 @@
+import sys
+from gevent.hub import PYGTE279
 from greentest import walk_modules, BaseTestCase, main
 import six
 
@@ -21,6 +23,12 @@ def make_exec_test(path, module):
 
 
 for path, module in walk_modules():
+    if sys.version_info[0] == 3 and path.endswith('2.py'):
+        continue
+    if sys.version_info[0] == 2 and path.endswith('3.py'):
+        continue
+    if not PYGTE279 and path.endswith('279.py'):
+        continue
     make_exec_test(path, module)
